feat: extract misc blueprint - ALL routes extracted from app.py
Agent-Logs-Url: https://github.com/abartrim/sobs/sessions/7f2e73ed-16ab-4410-b69c-e81f8332bee7
