+++ templates/ai_help.html
@@ -138,7 +138,7 @@
                     <a href="{{ url_for('settings.view_agent_rules') }}">Agent Settings</a> — create and manage autonomous agent rules.
                 </li>
                 <li>
-                    <a href="{{ url_for('view_work_items') }}">Work Items</a> — GitHub issues created by agent runs.
+                    <a href="{{ url_for('misc.view_work_items') }}">Work Items</a> — GitHub issues created by agent runs.
                 </li>
             </ul>
         </div>
