+++ templates/settings_agents.html
@@ -21,7 +21,7 @@
     <a href="{{ url_for('settings.view_ai_settings') }}" class="btn btn-sm btn-outline-primary" title="AI Settings">
       <i class="bi bi-gear me-1"></i><span class="agents-btn-label settings-btn-label">AI Settings</span>
     </a>
-    <a href="{{ url_for('view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
+    <a href="{{ url_for('misc.view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
       <i class="bi bi-arrow-left me-1"></i><span class="agents-btn-label settings-btn-label">Back to Settings</span>
     </a>
     <a href="{{ url_for('settings_agents_help') }}" class="btn btn-sm btn-outline-secondary page-help-btn" title="Help">
