+++ templates/settings_agents_help.html
@@ -9,7 +9,7 @@
            title="Open Agent Rules">
             <i class="bi bi-lightning-charge me-1"></i><span class="settings-btn-label">Agent Rules</span>
         </a>
-        <a href="{{ url_for('view_settings') }}"
+        <a href="{{ url_for('misc.view_settings') }}"
            class="btn btn-sm btn-outline-secondary"
            title="Back to Settings">
             <i class="bi bi-arrow-left me-1"></i><span class="settings-btn-label">Back to Settings</span>
