+++ templates/settings_ai.html
@@ -34,7 +34,7 @@
 </style>{% endblock %}
 {% block content %}
 {% set settings_ai_header_actions %}
-  <a href="{{ url_for('view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
+  <a href="{{ url_for('misc.view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="settings-btn-label">Back to Settings</span>
   </a>
   <a href="{{ url_for('settings_ai_help') }}" class="btn btn-sm btn-outline-secondary page-help-btn" title="Help">
