+++ templates/settings_help.html
@@ -4,7 +4,7 @@
 {% block styles %}{{ super() }}{% endblock %}
 {% block content %}
     {% set settings_help_header_actions %}
-        <a href="{{ url_for('view_settings') }}"
+        <a href="{{ url_for('misc.view_settings') }}"
            class="btn btn-sm btn-outline-secondary"
            title="Back to Settings">
             <i class="bi bi-arrow-left me-1"></i><span class="settings-btn-label">Back to Settings</span>
