+++ templates/settings_kubernetes.html
@@ -5,7 +5,7 @@
 {% block content %}
     {% set settings_kubernetes_header_actions %}
         <a class="btn btn-sm btn-outline-secondary"
-           href="{{ url_for('view_settings') }}"
+           href="{{ url_for('misc.view_settings') }}"
            title="Back to Settings">
             <i class="bi bi-arrow-left me-1"></i><span class="settings-btn-label">Back to Settings</span>
         </a>
