+++ templates/settings_masking.html
@@ -4,7 +4,7 @@
 {% block styles %}{{ super() }}{% endblock %}
 {% block content %}
 {% set settings_masking_header_actions %}
-  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('view_settings') }}" title="Back to Settings">
+  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('misc.view_settings') }}" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="masking-btn-label settings-btn-label">Back to Settings</span>
   </a>
   <a class="btn btn-sm btn-outline-secondary page-help-btn" href="{{ url_for('masking_help') }}" title="Help">
