+++ templates/settings_mcp.html
@@ -84,7 +84,7 @@
 </style>{% endblock %}
 {% block content %}
 {% set mcp_header_actions %}
-  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('view_settings') }}" title="Back to Settings">
+  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('misc.view_settings') }}" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="mcp-btn-label">Back to Settings</span>
   </a>
 {% endset %}
