+++ templates/settings_notifications.html
@@ -187,7 +187,7 @@
 {% endblock %}
 {% block content %}
 {% set settings_notifications_header_actions %}
-  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('view_settings') }}" title="Back to Settings">
+  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('misc.view_settings') }}" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="notif-btn-label settings-btn-label">Back to Settings</span>
   </a>
   <a class="btn btn-sm btn-outline-secondary page-help-btn" href="{{ url_for('settings_notifications_help') }}" title="Help">
