+++ templates/settings_notifications_help.html
@@ -9,7 +9,7 @@
            title="Open Notifications Settings">
             <i class="bi bi-bell me-1"></i><span class="settings-btn-label">Notifications</span>
         </a>
-        <a href="{{ url_for('view_settings') }}"
+        <a href="{{ url_for('misc.view_settings') }}"
            class="btn btn-sm btn-outline-secondary"
            title="Back to Settings">
             <i class="bi bi-arrow-left me-1"></i><span class="settings-btn-label">Back to Settings</span>
