+++ templates/settings_repositories.html
@@ -151,7 +151,7 @@
   <button type="button" class="btn btn-sm btn-outline-success" title="Onboarding Wizard" onclick="window.__sobsOpenOnboardingWizard && window.__sobsOpenOnboardingWizard()">
     <i class="bi bi-rocket-takeoff me-1"></i><span class="repositories-btn-label settings-btn-label">Onboarding Wizard</span>
   </button>
-  <a href="{{ url_for('view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
+  <a href="{{ url_for('misc.view_settings') }}" class="btn btn-sm btn-outline-secondary" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="repositories-btn-label settings-btn-label">Back to Settings</span>
   </a>
   <a href="{{ url_for('settings_repositories_help') }}" class="btn btn-sm btn-outline-secondary page-help-btn" title="Help">
