+++ templates/settings_tags.html
@@ -12,7 +12,7 @@
 
 {% block content %}
 {% set settings_tags_header_actions %}
-  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('view_settings') }}" title="Back to Settings">
+  <a class="btn btn-sm btn-outline-secondary" href="{{ url_for('misc.view_settings') }}" title="Back to Settings">
     <i class="bi bi-arrow-left me-1"></i><span class="tags-btn-label settings-btn-label">Back to Settings</span>
   </a>
   <a class="btn btn-sm btn-outline-secondary page-help-btn" href="{{ url_for('settings_tags_help') }}" title="Help">
