+++ templates/setup_playbooks_help.html
@@ -19,7 +19,7 @@
             </p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('summary') }}"
+           href="{{ url_for('misc.summary') }}"
            title="Back to Summary">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Summary</span>
         </a>
