+++ templates/work_items_help.html
@@ -19,7 +19,7 @@
             </p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_work_items') }}"
+           href="{{ url_for('misc.view_work_items') }}"
            title="Back to Work Items">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Work Items</span>
         </a>
